refactor: merge fetchAll & search into single 'clubList' redux module
