+++ src/modules/index.ts
@@ -3,8 +3,7 @@ import { all } from 'redux-saga/effects';
 import login, { loginSaga } from './login';
 import signup, { signupSaga } from './signup';
 import userData, { userDataSaga, likeSaga } from './userData';
-import fetchAll, { fetchAllSaga } from './fetchAll';
-import search, { searchSaga } from './search';
+import clubList, { clubListSaga } from './clubList';
 import fetchSingle, { fetchSingleSaga } from './fetchSingle'
 import post, { postSaga } from './post';
 import logout, { logoutSaga } from './logout';
@@ -16,9 +15,8 @@ const rootReducer = combineReducers({
     logout,
     signup,
     userData,
-    fetchAll,
+    clubList,
     fetchSingle,
-    search,
     post,
     comments,
     comment
@@ -34,9 +32,8 @@ export function* rootSaga() {
         logoutSaga(),
         signupSaga(), 
         userDataSaga(), 
-        fetchAllSaga(), 
+        clubListSaga(),
         fetchSingleSaga(),
-        searchSaga(),
         postSaga(),
         likeSaga(),
         commentsSaga(),
