--- src/modules/search/index.ts
@@ -1,4 +1,0 @@
-export { default } from './reducer';
-export * from './actions';
-export * from './types';
-export * from './sagas';
